Updated chartMaker function
In previous versions, if the user did not select a month after pressing the Re-generate button, an error would occur and the application would stop working. Now, we prevent this error by not executing the code after the user presses the Re-genarate button and asking the user to select the month first if they didn't do so. In addition, we have fixed several bugs such as failure to display the month correctly in the chart and failure to handle exceptions where BMI records were not found.